# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrocerEase Accra - Grocery Delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  residentId: number;
  residentName?: string;
  residentPhone?: string;
  residentAddress?: string;
  vendorId?: number;
  vendorName?: string;
  riderId?: number;
  riderName?: string;
  items: OrderItem[];
  subtotal: number;
  serviceFee: number;
  deliveryFee: number;
  total: number;
  status: OrderStatus;
  paymentMethod: string;
  isSubscription: boolean;
  callOnly: boolean;
  callAccepted?: boolean;
  photoUrl?: string;
  deliveryPhotoUrl?: string;
  pickupDeadline?: Date;
  eta?: string;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}